apparently my switch statement wasnt working, reimplemneted Wills mutli Ternary
